+++ src/app/candidate_lists/extractors/full_candidate_list.rs
@@ -1,8 +1,6 @@
 use axum::extract::Path;
 
-use crate::{
-    AppError, app::extractor::request_extractor, candidate_lists::FullCandidateList, trans,
-};
+use crate::{AppError, app::request_extractor, candidate_lists::FullCandidateList, trans};
 
 use super::CandidateListPathParams;
