+++ src/app/middleware/health.rs
