+++ src/app/middleware/health_tests.rs
