+++ src/app/political_groups/extractors/political_group.rs
@@ -1,4 +1,4 @@
-use crate::{app::extractor::request_extractor, political_groups::PoliticalGroup};
+use crate::{app::request_extractor, political_groups::PoliticalGroup};
 
 request_extractor!(PoliticalGroup, |store, parts, state| {
     Ok(store.get_political_group())
